+++ continuousintegration/orchestration/tests/cta_admin.sh
@@ -266,10 +266,11 @@ test_header 'tape'
 
 # Tape (ta)
 test_start "tape" "ta" "--all"
-
+admin_cta pl add --pl phys1 --ma man --mo mod --npcs 3 --npds 4
+admin_cta ll ch --name ${lls[1]} --pl phys1
 # Set added tape to full so we can test reclaim.
 test_and_check_cmd "Adding tape 'V01008'" "${command}" "add" "-v V01008 --mt T10K500G --ve vendor -l ${lls[1]} -t ctasystest -f true --purchaseorder order1"\
-  "select(.vid==\"V01008\" and .mediaType==\"T10K500G\" and .logicalLibrary==\"${lls[1]}\" and .full==true and .purchaseOrder==\"order1\") | .vid"\
+  "select(.vid==\"V01008\" and .mediaType==\"T10K500G\" and .logicalLibrary==\"${lls[1]}\" and .physicalLibrary==\"phys1\" and .full==true and .purchaseOrder==\"order1\") | .vid"\
   "1" "adding tape 'V01008'" || exit 1
 test_and_check_cmd "Reclaiming tape 'V01008'" "${command}" "reclaim" "-v V01008"\
   "select(.vid==\"V01008\" and .mediaType==\"T10K500G\" and .logicalLibrary==\"${lls[1]}\" and .full==false) | .vid"\
@@ -281,8 +282,9 @@ test_and_check_cmd "Changing tape V01008 order to order2" "${command}" "ch" "-v 
   "select(.vid==\"V01008\" and .purchaseOrder==\"order2\") | .vid"\
   "1" "changing tape V01008 order" || exit 1
 test_command "Removing tape V01008" "${command}" "rm" "-v V01008" || exit 1
+admin_cta ll ch --name ${lls[1]} --pl ""
+admin_cta pl rm --pl phys1
 test_assert || exit 1
-
 
 # Tape File (tf)
 test_start "tape file" "tf" "-v ${vids[0]}"
@@ -329,8 +331,10 @@ test_assert|| exit 1
 
 # Drive (dr)
 test_start "drive" "dr"
+admin_cta pl add --pl phys1 --ma man --mo mod --npcs 3 --npds 4
+admin_cta ll ch --name ${lls[1]} --pl phys1
 test_and_check_cmd "Setting drive '${dr_names_down[0]}' to UP" "${command}" "up" "${dr_names_down[0]} -r 'cta-admin systest up'"\
-  "select(.logicalLibrary==\"${dr_names_down[0]}\" and .driveName==\"${dr_names_down[0]}\" and .driveStatus==\"UP\" and .reason==\"cta-admin systest up\") | .driveName"\
+  "select(.logicalLibrary==\"${dr_names_down[0]}\" and .physicalLibrary==\"phys1\" and .driveName==\"${dr_names_down[0]}\" and .driveStatus==\"UP\" and .reason==\"cta-admin systest up\") | .driveName"\
   "1" "setting drive \'${dr_names_down[0]}\' to up state"|| exit 1
 test_and_check_cmd "Setting drive '${dr_names_down[0]}' to DOWN" "${command}" "down" "${dr_names_down[0]} -r 'cta-admin systest down'"\
   "select(.logicalLibrary==\"${dr_names_down[0]}\" and .driveName==\"${dr_names_down[0]}\" and .driveStatus==\"DOWN\" and .reason==\"cta-admin systest down\") | .driveName"\
@@ -339,6 +343,8 @@ test_and_check_cmd "Changing drive \'${dr_names_down[0]}\' message" "${command}"
   "select(.logicalLibrary==\"${dr_names_down[0]}\" and .comment==\"cta-admin test ch\") | .driveName"\
   "1" "changing drive \'${dr_names_down[0]}\' comment" || exit 1
 test_command "Removing drive \'${dr_names_down[0]}\'" "${command}" "rm" "${dr_names_down[0]}" || exit 1
+admin_cta ll ch --name ${lls[1]} --pl ""
+admin_cta pl rm --pl phys1
 test_assert_false || exit 1